Move `build-registry` into legacy folder.
