+++ addon-test-support/legacy-0-6-x/build-registry.js
